Refactor nvim installer to work properly with stow

+++ installers/nvim.sh
@@ -1,17 +1,16 @@
 #!/bin/bash
 # jasmith79's install script for neovim. 
-# Assumes python3/pip3 already installed. For old ubuntu and 
-# friends assumes the ppa has already been added.
 
-SOURCE="${BASH_SOURCE[0]}"
-while [ -h "$SOURCE" ]; do # resolve $SOURCE until the file is no longer a symlink
-  NVIMDIR="$( cd -P "$( dirname "$SOURCE" )" >/dev/null && pwd )"
-  SOURCE="$(readlink "$SOURCE")"
-  [[ $SOURCE != /* ]] && SOURCE="$NVIMDIR/$SOURCE" # if $SOURCE was a relative symlink, we need to resolve it relative to the path where the symlink file was located
-done
-NVIMDIR="$( cd -P "$( dirname "$SOURCE" )" >/dev/null && pwd )"
+# This script is located inside a file named installers
+# inside my dotfiles dir which should be inside $HOME
+# for stow to work properly.
+#
+# This will grab the path to the dotfiles dir, quoting to
+# guard against spaces in paths, e.g. I set my user name to
+# "Jared Smith" or something.
+DOTFILES_DIR = "$(dirname "$(dirname "$(readlink -f "$0")")"
 
-# Determine current user
+# Determine current user name
 user=$(logname)
 who=$(whoami)
 if [[ $user = "" ]]; then
@@ -22,15 +21,29 @@ if [[ $user = "" ]]; then
   user="$who"
 fi
 
-mkdir -p ~/.vim/pack/jsmith/start
 mkdir -p ~/.old_configs
-mkdir -p ~/.config/nvim
 
 if ! command -v nvim > /dev/null; then
   if command -v brew > /dev/null; then
     sudo -u "$user" brew install neovim
   elif command -v apt > /dev/null; then
-    sudo apt install vim neovim -y
+    # Official repo version too old for some of my plugins
+    sudo add-apt-repository ppa:neovim-ppa/stable
+    sudo apt update && sudo apt install vim neovim -y
+    # Need python support for some stuff
+    sudo apt install python3-dev python3-pip
+  else
+    echo "Unrecognized platform, aborting vim install"
+    exit 1
+  fi
+fi
+
+# Double-check we have stow.
+if ! command -v stow > /dev/null; then
+  if command -v brew > /dev/null; then
+    sudo -u "$user" brew install stow
+  elif command -v apt > /dev/null; then
+    sudo apt-get update && sudo apt install stow -y
   else
     echo "Unrecognized platform, aborting vim install"
     exit 1
@@ -38,12 +51,14 @@ if ! command -v nvim > /dev/null; then
 fi
 
 # If it's an actual file we don't want to lose it
-if [ -f "~/.config/nvim/init.vim" ]; then
+if [[ ! -L "~/.config/nvim/init.vim" && -f "~/.config/nvim/init.vim" ]]; then
   cp ~/.config/nvim/init.vim ~/.old_configs/init.vim
 fi
 rm -f ~/.config/nvim/init.vim
 
-ln -s $NVIMDIR/init.vim ~/.config/nvim/init.vim
+pushd $DOTFILES_DIR
+stow nvim
+popd
 
 echo "done. Installing vim-plug..."
 # install vim-plug for neovim and update neovim to use it
@@ -54,5 +69,5 @@ else
   echo "Requirement satisfied. Skipping..."
 fi
 
-# Install python3 support for neovim
-sudo -u "$user" python3 -m pip install --user neovim
+# Install python3 client for neovim
+sudo -u "$user" python3 -m pip install --user pynvim
